Restructure on library and app subdirs using module aliases
