+++ package.json
@@ -34,6 +34,7 @@
     "lint-staged": "^15.4.3",
     "ts-jest": "^29.2.5",
     "typescript": "^5.7.3",
-    "vite": "^6.1.0"
+    "vite": "^6.1.0",
+    "vite-tsconfig-paths": "^5.1.4"
   }
 }
