+++ src/bedrock/utils/mathUtils.test.ts
