+++ src/bedrock/utils/mathUtils.ts
