+++ src/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
       <title>App</title>
     <link rel="stylesheet" href="styles.css" />
-    <script type="module" src="app.ts"></script>
+    <script type="module" src="playground/app.ts"></script>
   </head>
   <body>
     <canvas id="display" />
