+++ vite.config.js
@@ -1,4 +1,6 @@
 import { defineConfig } from "vite";
+import tsconfigPaths from 'vite-tsconfig-paths';
+
 
 // https://vite.dev/config/
 export default defineConfig({
@@ -11,4 +13,5 @@ export default defineConfig({
     port: 8888,
     open: true,
   },
+  plugins: [tsconfigPaths()],
 });
